Add sketch of rake task to release via heroku api
This code does not run *yet* :cat: But its the basic idea. I need to
figure out what to do about the tarball on S3. If it is publically
accessible then whats all this work for? If its private then how do I
give the heroku-api access to it? S3 does not support basic auth.